🍋‍🟩 [FIX] revision for local test and ci update
