+++ docker-compose.yml
@@ -45,6 +45,7 @@ services:
     volumes:
       - ./frontend:/app
       - frontend_node_modules:/app/node_modules  # preserve installed modules
+      - playwright_cache:/root/.cache/ms-playwright  # persist browser binaries
     ports:
       - "3000:3000"
     env_file:
@@ -119,3 +120,4 @@ volumes:
   postgres_data:
   pgadmin_data:
   frontend_node_modules:
+  playwright_cache:
