+++ frontend/playwright.config.ts
@@ -1,14 +1,14 @@
 import { defineConfig, devices } from '@playwright/test'
 
 // Use GitHub Actions secret when available; fall back to fixed local test secret.
-// Must match the value used in e2e/global-setup.ts.
+// Must match the value used in tests/integration/global-setup.ts.
 const NEXTAUTH_SECRET =
   process.env.NEXTAUTH_SECRET || 'e2e-nextauth-secret-local-testing-only'
 
 export default defineConfig({
-  testDir: './e2e',
+  testDir: './tests/integration',
   timeout: 30_000,
-  globalSetup: './e2e/global-setup.ts',
+  globalSetup: './tests/integration/global-setup.ts',
   webServer: {
     command: 'npm run build && npm run start',
     url: 'http://localhost:3000',
@@ -22,7 +22,7 @@ export default defineConfig({
   },
   use: {
     baseURL: 'http://localhost:3000',
-    storageState: 'e2e/fixtures/auth-state.json',
+    storageState: 'tests/integration/fixtures/auth-state.json',
     screenshot: 'only-on-failure',
     video: 'retain-on-failure',
   },
