+++ frontend/tests/unit/filter-bar.test.tsx
@@ -1,8 +1,8 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 import { render, screen, fireEvent, waitFor } from '@testing-library/react'
-import { fetchArticleFilterSources, fetchArticleFilterTags } from '../lib/api/articles'
+import { fetchArticleFilterSources, fetchArticleFilterTags } from '@/lib/api/articles'
 
-vi.mock('../lib/api/articles', () => ({
+vi.mock('@/lib/api/articles', () => ({
   fetchArticleFilterSources: vi.fn(),
   fetchArticleFilterTags: vi.fn(),
 }))
@@ -55,13 +55,13 @@ describe('FilterBar', () => {
   })
 
   it('"Filters" toggle button is always rendered', async () => {
-    const { FilterBar } = await import('../components/features/articles/filter-bar')
+    const { FilterBar } = await import('@/components/features/articles/filter-bar')
     render(<FilterBar {...defaultProps} />)
     expect(screen.getByRole('button', { name: /filters/i })).toBeInTheDocument()
   })
 
   it('clicking "Filters" reveals Source and Tag popover triggers', async () => {
-    const { FilterBar } = await import('../components/features/articles/filter-bar')
+    const { FilterBar } = await import('@/components/features/articles/filter-bar')
     render(<FilterBar {...defaultProps} />)
     fireEvent.click(screen.getByRole('button', { name: /filters/i }))
     expect(screen.getByRole('button', { name: /source/i })).toBeInTheDocument()
@@ -69,14 +69,14 @@ describe('FilterBar', () => {
   })
 
   it('"Clear" button is hidden when activeFilterCount is 0', async () => {
-    const { FilterBar } = await import('../components/features/articles/filter-bar')
+    const { FilterBar } = await import('@/components/features/articles/filter-bar')
     render(<FilterBar {...defaultProps} />)
     fireEvent.click(screen.getByRole('button', { name: /filters/i }))
     expect(screen.queryByRole('button', { name: /clear/i })).not.toBeInTheDocument()
   })
 
   it('"Clear" button is visible when activeFilterCount > 0', async () => {
-    const { FilterBar } = await import('../components/features/articles/filter-bar')
+    const { FilterBar } = await import('@/components/features/articles/filter-bar')
     render(<FilterBar {...defaultProps} activeFilterCount={2} />)
     fireEvent.click(screen.getByRole('button', { name: /filters/i }))
     expect(screen.getByRole('button', { name: /clear/i })).toBeInTheDocument()
@@ -84,7 +84,7 @@ describe('FilterBar', () => {
 
   it('clicking Apply calls onApply with current draft state', async () => {
     const onApply = vi.fn()
-    const { FilterBar } = await import('../components/features/articles/filter-bar')
+    const { FilterBar } = await import('@/components/features/articles/filter-bar')
     render(<FilterBar {...defaultProps} sources={['rss']} onApply={onApply} activeFilterCount={1} />)
     fireEvent.click(screen.getByRole('button', { name: /filters/i }))
     fireEvent.click(screen.getByRole('button', { name: /apply/i }))
@@ -93,7 +93,7 @@ describe('FilterBar', () => {
 
   it('"Clear" resets filters and calls onApply with empty values', async () => {
     const onApply = vi.fn()
-    const { FilterBar } = await import('../components/features/articles/filter-bar')
+    const { FilterBar } = await import('@/components/features/articles/filter-bar')
     render(<FilterBar {...defaultProps} sources={['rss']} onApply={onApply} activeFilterCount={1} />)
     fireEvent.click(screen.getByRole('button', { name: /filters/i }))
     fireEvent.click(screen.getByRole('button', { name: /clear/i }))
@@ -103,7 +103,7 @@ describe('FilterBar', () => {
   })
 
   it('fetches source and tag options on mount', async () => {
-    const { FilterBar } = await import('../components/features/articles/filter-bar')
+    const { FilterBar } = await import('@/components/features/articles/filter-bar')
     render(<FilterBar {...defaultProps} />)
     await waitFor(() => {
       expect(fetchArticleFilterSources).toHaveBeenCalled()
