+++ frontend/tests/unit/scraper-source-form.test.tsx
@@ -5,21 +5,21 @@ describe('ScraperSourceForm', () => {
   beforeEach(() => vi.clearAllMocks())
 
   it('renders Source Type and Frequency fields', async () => {
-    const { ScraperSourceForm } = await import('../components/features/scraper/scraper-source-form')
+    const { ScraperSourceForm } = await import('@/components/features/scraper/scraper-source-form')
     render(<ScraperSourceForm onSubmit={vi.fn()} />)
     expect(screen.getByLabelText(/source type/i)).toBeInTheDocument()
     expect(screen.getByLabelText(/frequency/i)).toBeInTheDocument()
   })
 
   it('default source_type is rss', async () => {
-    const { ScraperSourceForm } = await import('../components/features/scraper/scraper-source-form')
+    const { ScraperSourceForm } = await import('@/components/features/scraper/scraper-source-form')
     render(<ScraperSourceForm onSubmit={vi.fn()} />)
     const select = screen.getByLabelText(/source type/i) as HTMLSelectElement
     expect(select.value).toBe('rss')
   })
 
   it('changing to blog reveals CSS selector fields', async () => {
-    const { ScraperSourceForm } = await import('../components/features/scraper/scraper-source-form')
+    const { ScraperSourceForm } = await import('@/components/features/scraper/scraper-source-form')
     render(<ScraperSourceForm onSubmit={vi.fn()} />)
     fireEvent.change(screen.getByLabelText(/source type/i), { target: { value: 'blog' } })
     expect(screen.getByLabelText(/article link/i)).toBeInTheDocument()
@@ -28,7 +28,7 @@ describe('ScraperSourceForm', () => {
   })
 
   it('changing back to rss hides selector fields', async () => {
-    const { ScraperSourceForm } = await import('../components/features/scraper/scraper-source-form')
+    const { ScraperSourceForm } = await import('@/components/features/scraper/scraper-source-form')
     render(<ScraperSourceForm onSubmit={vi.fn()} />)
     fireEvent.change(screen.getByLabelText(/source type/i), { target: { value: 'blog' } })
     fireEvent.change(screen.getByLabelText(/source type/i), { target: { value: 'rss' } })
@@ -37,7 +37,7 @@ describe('ScraperSourceForm', () => {
 
   it('submit with RSS data calls onSubmit with correct shape', async () => {
     const onSubmit = vi.fn()
-    const { ScraperSourceForm } = await import('../components/features/scraper/scraper-source-form')
+    const { ScraperSourceForm } = await import('@/components/features/scraper/scraper-source-form')
     render(<ScraperSourceForm onSubmit={onSubmit} />)
     fireEvent.change(screen.getByLabelText(/name/i), { target: { value: 'HN' } })
     fireEvent.change(screen.getByLabelText(/url/i), { target: { value: 'https://hn.com/rss' } })
@@ -49,7 +49,7 @@ describe('ScraperSourceForm', () => {
 
   it('submit with blog data includes selector_config', async () => {
     const onSubmit = vi.fn()
-    const { ScraperSourceForm } = await import('../components/features/scraper/scraper-source-form')
+    const { ScraperSourceForm } = await import('@/components/features/scraper/scraper-source-form')
     render(<ScraperSourceForm onSubmit={onSubmit} />)
     fireEvent.change(screen.getByLabelText(/source type/i), { target: { value: 'blog' } })
     fireEvent.change(screen.getByLabelText(/article link/i), { target: { value: 'a.post' } })
